requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-anthropic==0.1.1
plotly==5.18.0
pandas==2.1.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
anthropic==0.18.0
pyyaml==6.0.1